chore: upgrade dev dependencies and split CodeMirror chunk
Upgrade vite 6→8, typescript 5→6, @vitejs/plugin-react 4→6, turbo 2.4→2.9,
eslint 10.1→10.2, @types/node 22→25 to keep dev tooling current.

Migration changes required by the upgrades:
- tsconfig: add types: [\"bun\"] (root) and [\"bun\", \"vite/client\"] (webui)
  to opt back into type packages now that TS 6 defaults types to []
- google-cache.ts: format caught error via instanceof check for the new
  @typescript-eslint/restrict-template-expressions strictness
- ProjectPage: move panelWidthRef.current write into useEffect to satisfy
  the new react-hooks/refs rule (React Compiler compatibility)
- SettingsView: drop unused CustomProviderDef import, mark fire-and-forget
  handleDeleteProvider with void
- eslint: bump react-hooks preset to recommended-latest

Additionally split the Library page into a lazy chunk and force CodeMirror
into its own chunk via Rolldown codeSplitting groups. Main bundle drops
from 1031 KB to 521 KB (gzip: 338 → 160 KB) — 178 KB gzip saved on first
paint, since CodeMirror is only fetched when entering the Library page.

Required for the split to actually work:
- apps/webui/package.json: declare sideEffects: [\"**/*.css\"] so the
  bundler can tree-shake non-CSS modules through the shared/ui barrel
- shared/ui/index.ts: drop TextEditor re-export so CodeMirror is
  unreachable from the static graph; SkillEditor and LibraryEditorView
  now deep-import TextEditor directly
- vite.config.ts: add codeSplitting group for @codemirror / @lezer
