+++ apps/webui/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "license": "MIT",
   "type": "module",
+  "sideEffects": ["**/*.css"],
   "scripts": {
     "dev": "bun scripts/dev.ts",
     "dev:server": "bun --watch src/server/index.ts",
@@ -36,11 +37,11 @@
     "@tailwindcss/vite": "^4.1.0",
     "@types/react": "^19.0.0",
     "@types/react-dom": "^19.0.0",
-    "@vitejs/plugin-react": "^4.4.0",
+    "@vitejs/plugin-react": "^6.0.1",
     "babel-plugin-react-compiler": "^1.0.0",
     "sharp": "^0.34.5",
     "tailwindcss": "^4.1.0",
-    "typescript": "^5.7.0",
-    "vite": "^6.3.0"
+    "typescript": "^6.0.2",
+    "vite": "^8.0.5"
   }
 }
