+++ apps/webui/src/client/shared/ui/index.ts
@@ -1,7 +1,9 @@
 export { Button } from "./Button.js";
 export { Dialog } from "./Dialog.js";
 export { Badge } from "./Badge.js";
-export { TextEditor, type EditorLanguage } from "./TextEditor.js";
+// TextEditor (and its EditorLanguage type) is intentionally NOT re-exported here
+// — it pulls in CodeMirror. Import it directly from "./TextEditor.js" so it
+// stays in the lazy LibraryPage chunk.
 export { Select } from "./Select.js";
 export { FormField } from "./FormField.js";
 export { TabBar } from "./TabBar.js";
