+++ apps/webui/tsconfig.json
@@ -14,6 +14,7 @@
     "declaration": true,
     "declarationMap": true,
     "sourceMap": true,
+    "types": ["bun", "vite/client"],
     "paths": {
       "@/*": ["./src/*"]
     }
