+++ package.json
@@ -20,11 +20,11 @@
   "devDependencies": {
     "@eslint/js": "^10.0.1",
     "@types/bun": "^1.3.11",
-    "@types/node": "^22.0.0",
-    "eslint": "^10.1.0",
+    "@types/node": "^25.5.2",
+    "eslint": "^10.2.0",
     "eslint-plugin-react-hooks": "^7.0.1",
-    "turbo": "^2.4.0",
-    "typescript": "^5.7.0",
+    "turbo": "^2.9.4",
+    "typescript": "^6.0.2",
     "typescript-eslint": "^8.58.0"
   },
   "packageManager": "bun@1.3.11"
