+++ tsconfig.json
@@ -12,7 +12,8 @@
     "declarationMap": true,
     "sourceMap": true,
     "isolatedModules": true,
-    "verbatimModuleSyntax": true
+    "verbatimModuleSyntax": true,
+    "types": ["bun"]
   },
   "exclude": ["node_modules", "dist"]
 }
